Доделал функционал ролей. Использовал роли для редактирования категорий

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -27,6 +27,8 @@ class CategoryController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Category::class);
+
         return view('admin.categories.create');
     }
 
@@ -38,6 +40,8 @@ class CategoryController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', Category::class);
+
         $params = $request->all();
         unset($params['image']);
         if ($request->has('image')) {
@@ -66,6 +70,8 @@ class CategoryController extends Controller
      */
     public function edit(Category $category)
     {
+        $this->authorize('update', $category);
+
         return view('admin.categories.edit', compact('category'));
     }
 
@@ -78,6 +84,8 @@ class CategoryController extends Controller
      */
     public function update(Request $request, Category $category)
     {
+        $this->authorize('update', $category);
+
         $params = $request->all();
         unset($params['image']);
         if ($request->has('image')) {
@@ -96,6 +104,8 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
+        $this->authorize('delete', $category);
+
         $category->delete();
         Storage::delete($category->image);
         return redirect()->route('admin.categories.index');
